requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=0.25.0=pypi_0
addict=2.4.0=pypi_0
aenum=3.1.15=pypi_0
aiofiles=23.2.1=pypi_0
aiohttp=3.9.3=pypi_0
aiosignal=1.3.1=pypi_0
aliyun-python-sdk-core=2.14.0=pypi_0
aliyun-python-sdk-kms=2.16.2=pypi_0
altair=5.2.0=pypi_0
annotated-types=0.6.0=pypi_0
anyio=4.3.0=pypi_0
archspec=0.2.3=pyhd3eb1b0_0
asttokens=2.4.1=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
bitsandbytes=0.42.0=pypi_0
blas=1.0=openblas
bleach=6.1.0=pypi_0
boltons=23.0.0=py39h06a4308_0
brotli-python=1.0.9=py39h6a678d5_7
bzip2=1.0.8=h5eee18b_5
c-ares=1.19.1=h5eee18b_0
ca-certificates=2023.12.12=h06a4308_0
certifi=2024.2.2=py39h06a4308_0
cffi=1.16.0=py39h5eee18b_0
chardet=5.2.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
clip=1.0=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.1=pypi_0
compel=2.0.2=pypi_0
compound-word-splitter=0.4=pypi_0
conda=24.1.2=py39h06a4308_0
conda-libmamba-solver=24.1.0=pyhd3eb1b0_0
conda-package-handling=2.2.0=py39h06a4308_0
conda-package-streaming=0.9.0=py39h06a4308_0
contourpy=1.2.0=pypi_0
crcmod=1.7=pypi_0
cryptography=42.0.5=pypi_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.12.1=pypi_0
debugpy=1.8.1=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
diffusers=0.28.0.dev0=dev_0
dill=0.3.8=pypi_0
distro=1.8.0=py39h06a4308_0
docopt=0.6.2=pypi_0
einops=0.7.0=pypi_0
enchant=0.0.1=pypi_0
exceptiongroup=1.2.0=pypi_0
executing=2.0.1=pypi_0
fastapi=0.110.0=pypi_0
fastjsonschema=2.19.1=pypi_0
fastsam=0.0.1.dev0=pypi_0
ffmpy=0.3.2=pypi_0
filelock=3.13.1=pypi_0
fmt=9.1.0=hdb19cb5_0
fonttools=4.49.0=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.2.0=pypi_0
ftfy=6.1.3=pypi_0
future=0.18.3=py39h06a4308_0
gdown=5.1.0=pypi_0
gradio=3.35.2=pypi_0
gradio-client=0.10.1=pypi_0
grpcio=1.62.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.4=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.21.3=pypi_0
icu=73.1=h6a678d5_0
idna=3.6=pypi_0
ignite=1.1.0=pypi_0
imageio=2.34.0=pypi_0
importlib-metadata=7.0.1=pypi_0
importlib-resources=6.1.2=pypi_0
ipykernel=6.29.3=pypi_0
ipython=8.12.3=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.3=pypi_0
jmespath=0.10.0=pypi_0
jsonpatch=1.32=pyhd3eb1b0_0
jsonpointer=2.1=pyhd3eb1b0_0
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter-client=8.6.0=pypi_0
jupyter-core=5.7.1=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
kiwisolver=1.4.5=pypi_0
krb5=1.20.1=h143b758_1
lazy-loader=0.3=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libarchive=3.6.2=h6ac8c49_2
libcurl=8.5.0=h251f7ec_0
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libmamba=1.5.6=haf1ee3a_0
libmambapy=1.5.6=py39h2dafd23_0
libnghttp2=1.57.0=h2d74bed_0
libopenblas=0.3.21=h043d6bf_0
libprotobuf=3.20.3=he621ea3_0
libsolv=0.7.24=he621ea3_0
libssh2=1.10.0=hdbd6064_2
libstdcxx-ng=11.2.0=h1234567_1
libxml2=2.10.4=hf1b16e4_1
lightning-utilities=0.11.0=pypi_0
linkify-it-py=2.0.3=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.5.2=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
mat4py=0.6.0=pypi_0
matplotlib=3.8.3=pypi_0
matplotlib-inline=0.1.6=pypi_0
mdit-py-plugins=0.3.3=pypi_0
mdurl=0.1.2=pypi_0
menuinst=2.0.2=py39h06a4308_0
mistune=3.0.2=pypi_0
mmcv=2.1.0=pypi_0
mmdeploy=1.3.1=pypi_0
mmdet=3.3.0=dev_0
mmengine=0.10.3=pypi_0
mmpretrain=1.2.0=dev_0
mmsegmentation=1.2.2=pypi_0
model-index=0.1.11=pypi_0
modelindex=0.0.2=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
multiprocess=0.70.16=pypi_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.3=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
networkx=3.2.1=pypi_0
ninja=1.11.1.1=pypi_0
ninja-base=1.10.2=hd09550d_5
numpy=1.26.4=py39heeff2f4_0
numpy-base=1.26.4=py39h8a23956_0
nvidia-cublas-cu11=11.11.3.6=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu11=11.8.87=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu11=11.8.89=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu11=11.8.89=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu11=8.7.0.84=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu11=10.3.0.86=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu11=11.4.1.48=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu11=11.7.5.86=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu11=2.19.3=pypi_0
nvidia-nccl-cu12=2.19.3=pypi_0
nvidia-nvjitlink-cu12=12.3.101=pypi_0
nvidia-nvtx-cu11=11.8.86=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
onnx=1.15.0=pypi_0
open-clip-torch=2.24.0=pypi_0
openai=1.16.2=pypi_0
opencv-python=4.9.0.80=pypi_0
opendatalab=0.0.10=pypi_0
openmim=0.3.9=pypi_0
openssl=3.0.13=h7f8727e_0
openxlab=0.0.34=pypi_0
ordered-set=4.1.0=pypi_0
orjson=3.9.15=pypi_0
oss2=2.17.0=pypi_0
ovam=0.0.1=dev_0
packaging=23.2=pypi_0
pandas=2.2.1=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.3=pypi_0
pcre2=10.42=hebb0a14_0
pexpect=4.9.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=10.2.0=pypi_0
pip=19.3.1=pypi_0
pipreqs=0.5.0=pypi_0
platformdirs=3.10.0=py39h06a4308_0
pluggy=1.0.0=py39h06a4308_1
prettytable=3.10.0=pypi_0
prompt-toolkit=3.0.43=pypi_0
protobuf=3.20.2=pypi_0
psutil=5.9.8=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pybind11-abi=4=hd3eb1b0_1
pycocotools=2.0.7=pypi_0
pycosat=0.6.6=py39h5eee18b_0
pycparser=2.21=pyhd3eb1b0_0
pycryptodome=3.20.0=pypi_0
pydantic=2.6.3=pypi_0
pydantic-core=2.16.3=pypi_0
pydub=0.25.1=pypi_0
pygments=2.17.2=pypi_0
pyparsing=3.1.1=pypi_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.18=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
python-multipart=0.0.9=pypi_0
pytorch-fid=0.3.0=pypi_0
pytorch-ignite=0.4.13=pypi_0
pytz=2023.4=pypi_0
pyyaml=6.0.1=py39h5eee18b_0
pyzmq=25.1.2=pypi_0
readline=8.2=h5eee18b_0
referencing=0.33.0=pypi_0
regex=2023.12.25=pypi_0
reproc=14.2.4=h295c915_1
reproc-cpp=14.2.4=h295c915_1
requests=2.28.2=pypi_0
rich=13.4.2=pypi_0
rpds-py=0.18.0=pypi_0
ruamel.yaml=0.17.21=py39h5eee18b_0
ruamel.yaml.clib=0.2.6=py39h5eee18b_1
safetensors=0.4.2=pypi_0
scikit-image=0.22.0=pypi_0
scikit-learn=1.4.1.post1=pypi_0
scipy=1.12.0=pypi_0
seaborn=0.13.2=pypi_0
segment-anything=1.0=pypi_0
semantic-version=2.10.0=pypi_0
sentencepiece=0.2.0=pypi_0
setuptools=60.2.0=pypi_0
shapely=2.0.3=pypi_0
simplecrf=0.2.1.1=pypi_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.3=pypi_0
starlette=0.36.3=pypi_0
sympy=1.12=pypi_0
tabulate=0.9.0=pypi_0
termcolor=2.4.0=pypi_0
terminaltables=3.1.10=pypi_0
threadpoolctl=3.3.0=pypi_0
tifffile=2024.2.12=pypi_0
timm=0.9.16=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.15.2=pypi_0
tomli=2.0.1=pypi_0
toolz=0.12.1=pypi_0
torch=2.2.1=pypi_0
torch-fidelity=0.3.0=pypi_0
torchaudio=2.2.1+cu118=pypi_0
torchmetrics=1.3.2=pypi_0
torchvision=0.17.1+cu118=pypi_0
tornado=6.4=pypi_0
tqdm=4.66.2=pypi_0
traitlets=5.14.1=pypi_0
transformers=4.36.0=pypi_0
triton=2.2.0=pypi_0
typing-extensions=4.10.0=pypi_0
typing_extensions=4.9.0=py39h06a4308_1
tzdata=2024.1=pypi_0
uc-micro-py=1.0.3=pypi_0
ultralytics=8.0.120=pypi_0
urllib3=1.26.18=pypi_0
uvicorn=0.27.1=pypi_0
wcwidth=0.2.13=pypi_0
webencodings=0.5.1=pypi_0
websockets=12.0=pypi_0
wheel=0.41.2=py39h06a4308_0
wrapt=1.16.0=pypi_0
xz=5.4.6=h5eee18b_0
yaml=0.2.5=h7b6447c_0
yaml-cpp=0.8.0=h6a678d5_0
yapf=0.40.2=pypi_0
yarg=0.1.9=pypi_0
yarl=1.9.4=pypi_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstandard=0.19.0=py39h5eee18b_0
zstd=1.5.5=hc292b87_0
